ART: Fix debug line info for method entry code
DWARF debug_line info for a method started with an entry
correspondent to the first dex opcode. Thus the method prolog,
which preceded this entry, was mapped to the previous method.

This fix is to change the address of the first line table
entry to the method's start address. This makes both the method's
prolog and the following block mapped to the method's source line.

In the following example the address 0x101c mapped to line 22.
The fix makes the address 0x101c map to line 33.

Before (2 entries):
   0x1010 line 22
   // 0x101c the next method starts here
   0x101f line 33

After (3 entries):
   0x1010 line 22
   0x101c line 33 // the next method starts here
   0x101f line 33

Signed-off-by: Yevgeny Rouban <yevgeny.y.rouban@intel.com>
Change-Id: If2079d00083af3ae2e1d7daee36e9336c6df1c29