[optimizing][reference] Support temp register pair
This is a reference patch and should not be merged at this point.

The X86 intrinsics work at one point needed a temporary register pair to
implement long min/max, as well as the 2 input register pairs.  I
attempted to use 2 single registers, but the register allocator failed
because the 2 temporary singles were from 2 different register pairs,
and there were no longer 3 register pairs available, which is all that
there are on X86.  This patch adds support for allocating a register
pair as a temporary, which ensured that the 2 registers were
consecutive.

I was able to implement the intrinsic with only 1 register, so this
code wasn't needed after all.  This records the change in case this is
needed at a later poing.

Change-Id: I1d0f24d303dcb838106482596930c0d3ab49d09b
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>