RFC: Add explicit use of current method in InvokeStaticOrDirect.
Not ready to submit yet. Just want to get some feedback whether the
calling convention will be kept the same in the future.

This patch adds a new instruction HCurrentMethod so that we can allocate
a register for it.

Direct call before:         After:
  ldr w0, [sp]
  ldr w0, [x0, #12]           ldr w0, [x20, #12]
  ldr w0, [x0, #16]           ldr w0, [x0, #16]
  ldr lr, [x0, #56]           ldr lr, [x0, #56]
  blr lr                      blr lr

This patch is not completed yet. There are still other instructions which
implicitly use current method reference. Also slowpaths have not been
updated.

Change-Id: If8c2945bbabae771c5ddcbe9ea1ad1a9890a464a