ART: General-case negation in boolean simplifier
Code transformations may optimize out the condition instruction of
an If and replace it with a boolean value. In such case, the boolean
simplifier would not know how to negate the condition and would fail.
This patch implements negation in this generic case with 'equals 0'
as a substitute for the non-existing boolean NOT instruction.

Bug: 19992954
Change-Id: I152036fcc6bbecccc767d3024a5c060177597d88