Remove duplicates phis created during SSA transformation
Because of equality comparisons against null constants we may and up
with the duplicates phis (same type) for the same dex registers. This is
correct but generats more code and prevent some optimizations.

This CL adds another step in the SSA builder to remove the extra Phi
nodes created due to equality operators.

The graph checker checks that for a given dex register not two phis have
the same type.

We also replace zero int constant with null constant when we compare a
reference against null.

Change-Id: Id37cc11a016ea767c7e351575e003d822a9d2e60