Optimizing: Fix ParallelMoveResolver for cycles
ParallelMoveResolver should be careful with order of handling
swap for moves in 64-bit world as well as probably for FP where
one register is used for double. If there are two moves in a
cycle with different type sizes (32-bit and 64-bit) then wider
move should be swapped first otherwise we can lose the high
part of 64-bit move if 32-bit move use stack slot as location.

Unfortunately Location and MoveOperand do not contain information
about size of arguments and if register can be used to keep
both 64-bit and 32-bit values (like it happens in 64-bit) then
we can trouble to detect which move is wider.

The patch implements safe and best guess approach to determine
the order.

Change-Id: I9e89da07b270c626fa85ace9496fbb93412ab767
Signed-off-by: Serguei Katkov <serguei.i.katkov@intel.com>