[optimizing] Add support for switch jump tables
Implement the HSwitch HInstruction to generate switch tables rather than
chained if statements.  Implement this for x86_64 with RIP support,
since it has a constant table.

Add a virtual function to CodeGenerator to allow a backend to implement
support for switch tables.

Add a pointer to the CodeGenerator to the HGraph, as the builder needs
to be able to query to see if HSwitch should be generated.

Change-Id: I3c53a447830452fc514d1c7b92ac9d8c8d785d6e
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>