[optimizing] Add X86 support of switch jump tables
Implement the HSwitch HInstruction to generate switch tables rather than
chained if statements.  Implement this for x86_64 with RIP support,
since it has a constant table.  For x86, we need to access the constant
area, so invent an HX86Switch instruction to be used.  Since the
baseline compiler doesn't have access to a constant area, HSwitch needs
to be implemented as compare/jump pairs as well.

Handle dead code elimination of Switch properly.

Change-Id: I3c53a447830452fc514d1c7b92ac9d8c8d785d6e
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>