[WIP][optimizing] Enable inlining of throwers under special condition
The SpecialMethodInliner in Quick can handle inlining of throwing
getters if they are synthetic accessors. What this means is that Quick's
inliner can handle cases which Optimizing compiler cannot handle.

In an effort to mitigate that, the Optimizing inliner has been enhanced
to accept throwers when debugging is disabled. The technique used is
more general than the one in Quick so now this represents a superset
of methods inlined by Quick. The allowed throwers are NullCheck,
BoundCheck, and DivZeroCheck.

TODO: Still need to add tests.

Change-Id: I433e78a5f2b965feddbc4de4f38a26febf9a79c7
Signed-off-by: Razvan A Lupusoru <razvan.a.lupusoru@intel.com>