Stop building fuzzer by default.
Fuzzer depends now on the host core image and forcing this dependency
when building (mm) slows down development (if you mess up dex2oat build
somehow you need to first revert to a stable state to build the core
image to be able to move foreward). Since this is a tool, it is ok to
build it only on demand.

Change-Id: I8dd6d12cf8d43751af5eb53eb466e5851455c591