Remove fuzzer build-time dependency on core image.
Test if the image exists at runtime.

Forcing this dependency slows down development. For example if you mess
up dex2oat build somehow you need to first revert to a stable state to
build the core image to be able to move foreward. Since this is a test
tool, it is ok to assume that the core image has already been build if
host execution was selected.

Change-Id: Id61a960ba62bb8d1d21acb07114f72208fbc9720