[WIP][optimizing] Fuse compare/if (x86_64 first)
This is a preliminary implementation of fusing long/float/double
compares with ifs to avoid materializing the result from the compare.

It is implemented for x86_64 at this point, as it is using a CL that
adds a CodeGenerator pointer to the HGraph.

The simplifier is used to replace a HIf/HCondition/HCompare with a
HCompareIf if it is valid to do so.

HELP NEEDED:  I don't know how to delete the existing
HCompare/HCondition.  The attempts I tried failed because the
Uses/EnviromentUses were not empty.  I am also not sure if the
Environment is properly maintained.  Please advise.

Change-Id: I3112dd3c19ca2a0fdc67e88f0b587424720ebdd2
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>