JDWP: fix breakpoint on method entry when tracing is enabled
When tracing is enabled, there is always a listener for MethodEntry
event. However, if the debugger is only listening to DexPcMoved
event (to manage JDWP Breakpoint event), it won't be notified
and we won't report the breakpoint to the debugger.

This CL adds the Thread::tls_32bit_sized_values::debug_method_entry_
flag to give more control to the debugger to properly manage the
MethodEntry event. This flag replaces the local variable
notified_method_entry_event in the interpreter and simplifies the
code.

Bump oat version to force recompilation because the layout of the
Thread class is modified.

Bug: 19829329
Change-Id: I204af9112e37d2eebc86661fb7c961a41c74e598