[optimizing] Fuse compare/if (x86/x64-64 first)
This is a preliminary implementation of fusing long/float/double
compares with conditions and ifs to avoid materializing the result
from the compare and condition.

A HCompare as not needing materialization, if it is safe to do so.
This is done in a manner such that a backend that doesn't know how
to handle this case will continue to function as is.

There is a restriction that the Compare/Condition/If must be
consecutive (ignoring parallel moves), in order to ensure that there is
no suspend point where it would be possible to tell that the result of
the Compare is not stored in a vreg.   There must also be only one use
of the compare result and the condition result, and no environment uses.

Change-Id: I1f1db206e6dc336270cd71070ed3232dedc754d6
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>