[WIP][opt] Fuse compare/condition (x86/x64-64)
This is a preliminary implementation of fusing long/float/double
compares with conditions to avoid materializing the result from the
compare and condition.

The information from a HCompare is transferred to the HCondition if it
is legal.  There must be only a single use of the HCompare, the HCompare
and HCondition must be in the same block, the HCondition must not need
materialization.

NOTE: This is implemented only for X86 and X86_64 at this time.
This needs to be implemented for ARM & ARM64 as well before it can be
merged.

Added GetOppositeCondition() to HCondition to return the flipped
condition.

Change-Id: I1f1db206e6dc336270cd71070ed3232dedc754d6
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>