Revert "Revert "[optimizing] Improve x86 shifts""
This reverts commit f9aac1e9f442c2486cd54f045d43e15791601205.

Ensure that when generating baseline code, we don't use Location::Any()
for the first input if the output is Location::SameAsFirstInput().
We must use RequiresRegister() to ensure that an input VR on the stack
won't be inadvertently changed as a side effect of the instruction.

This is all due to the 2 operand instructions on X86.

Change-Id: I400834052b114abf0d616da1b4b6506f7bba10ab