Optimizing: Fix SsaBuilder
This CL changes SsaBuilder in two places.
The first one fixes PrimitiveTypePropagation. If we update
type of input in place we must re-consider its Phi users.
It might happen that it is not a loop and it was already considered.
Included test shows this issue.

The second one eliminates redundant Phi node. It is possible that
after primitive type propagation there are two identical Phi nodes.
One of them can be removed. The same test can be used to see
the issue and solution.

Change-Id: I0305c8aa0fde69ebdcf1ac5ff568aa5fbbc8a353
Signed-off-by: Serguei Katkov <serguei.i.katkov@intel.com>