[optimizing] Replace FP divide by power of 2
Replace a floating point division by a power of two by a multiplication
of the reciprocal.  This is guarenteed to have the exact same result as
it is exactly representable.

Add routines to allow generation of float and double constants after the
SSA Builder.  I was unsure if float and double caches should be
implement.  Under the assumption that there is probably not a lot of
repetition of FP values.  Please let me know.

Change-Id: I3a6c3847b49b4e747a7e7e8843ca32bb174b1584
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>