Remove duplicates phis created during SSA transformation
When creating equivalent phis we copy the inputs of the original phi
which be improperly typed. This will be fixed during the type
propagation butas a result we may have two equivalent phis with the same
type for the same dex register. This is correct but generats more code
and prevent some optimizations.

This CL adds another step in the SSA builder to remove the extra Phi
nodes created due to equality operators.

The graph checker verifies that for a given dex register not two phis
have the same type.

Also, replace zero int constant with null constant when we compare a
reference against null.

Change-Id: Id37cc11a016ea767c7e351575e003d822a9d2e60