[optimizing] Add support for x86 constant area
Use the Quick trick of finding the address of the method by calling the
next instruction and popping the return address into a register.  This
trick is used because of the lack of PC-relative addressing in 32 bit
mode on the X86.

Add an optional 'backend optimization' pass that is per code-generator
that can examine the graph and add instructions before register
allocation.  Add a HX86ComputeBaseMethodAddress instruction to trigger
generation of the method address, which is referenced by instructions
needing access to the constant area.

Add a 'backend input' field to HInstruction that allows the backend to
specify that this instruction needs an extra input (such as the method
address).  Treat this an extra input when needed.

When creating Locations, add the method address input as needed.  Code
generation will use this input to address the constant area.

Currently, this is used only for the basic FP operations.  Other uses
will be added later.

Port the X86_64 assembler constant area handling to the X86.

Use the new 'per-backend' infrastructure.

Rebased on MIPS64 fix to new 'per-backend' infrastructure.

Change-Id: I6d235a72238262e4f9ec0f3c88319a187f865932
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>