Remove ArtMethod* parameter from dex cache entry points.
Load the ArtMethod* from the stack instead. This reduces the
size of the generated code.

Three of the entry points are called only from a slow-path
and the fourth (InitializeTypeAndVerifyAccess) is rare and
already slow enough that the extra load from stack, in case
we have an ArtMethod* in register, is insignificant. And as
we're starting to use PC-relative addressing of the dex
cache arrays (already done by Quick for boot image), having
the ArtMethod* in a register becomes less likely anyway.

Change-Id: Ib19b9d204e355e13bf386662a8b158178bf8ad28