[optimizing] More x86_64 code gen improvements
Allow more operands for Add and Sub. Allow a constant length for
BoundCheck.  Can both the index and length of BoundsCheck be constants?

I attempted to allow 3 operand Sub if the RHS was a constant, but that
failed since the constant was already in a register, and the register
was supplied to the code generator rather than the constant.

Change-Id: I2c7adc6e733cf8ce6997aba76aa763d0835bd2d6
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>