Fix wrong deletion of save_destination when compiled with gcc 4.9.
When compiled with gcc 4.9 and default -O2 optimization -finline-small-functions,
the save_destination object was deleted more than once.
I am not sure yet if this is an incorrect optimization according to the standard
C++ or just some compiler-dependent behavior.

Anyway this patch will fix the problem and capture the save_destination pointer
instead of the smart pointer. Later we should simplify the many classes used by
parsed_options.cc to reduce inefficient and error-prone dynamic allocation of objects.

BUG: 19872411
Change-Id: I6bc803a38d1bdbb6b163d5d58372a349d078e94c