Increase frame size limit to enable Mips64 clang builds
Former limit of 1728 bytes was sufficient for all targets
when using gcc, and was sufficient for all targets except
Mips64 when using clang.  Mips64 builds using clang
succeed when using a 2% higher limit of 1760, and
fail when using a limit of 1759 or less.
Mips64 native stack frames are 16-byte aligned.

The nature of clang's increase over gcc here,
has not been investigated yet.

Change-Id: Id77ea44713d9bbcf4af20fd9252bd987ce578779