[WIP] Optimizing: Better invoke-static/-direct dispatch.
Add framework for different types of loading ArtMethod*
and code pointer retrieval. Implement invoke-static and
invoke-direct calls the same way as Quick.

PC-relative loads from dex cache arrays are used only for
x86-64 and arm64. The implementation for other architectures
will be done in separate CLs.

On arm64, we don't use relative calls for boot to boot calls
across dex files because the code generator doesn't know
that we are compiling a boot image. Instead we load the
code pointer from a literal pool.

Change-Id: I468ca4d422dbd14748e1ba6b45289f0d31734d94