Optimizing: Better invoke-static/-direct dispatch.
Add framework for different types of loading ArtMethod*
and code pointer retrieval. Implement invoke-static and
invoke-direct calls the same way as Quick.

PC-relative loads from dex cache arrays are used only for
x86-64 and arm64. The implementation for other architectures
will be done in separate CLs.

On arm64 we currently don't have an appropriate literal
support, so we don't deduplicate method address literals.
Additionally, for calls that would load the code address
from a literal pool in Quick, we load it from the ArtMethod.

Change-Id: I468ca4d422dbd14748e1ba6b45289f0d31734d94