ART: Stricter matching of Checker lines
This patch refactors the MatchLines function of Checker and changes it
so that regular expressions must match entire words of C1vis output.
Previously, this could lead to false positives because assertion 'xyz'
translated to 'xyz.*' and hence testing for id 'i12' could be satisfied
by the presence of 'i123'.

Change-Id: Iaeb486c53519b450ea0cd40042ff3048b38d5a7b