Fix premature deoptimization if the loop body isn't entered.
Add a test between initial_ and end_ to see if the loop body is entered.
If the loop body isn't entered at all, we jump to the loop header. Loop header is
still executed and is going to test the condition again and loop body won't be
entered. This makes sure no deoptimization is triggered if the loop body isn't
even entered. Instead of jumping to loop header directly, it needs to jump to the
instruction that follows the last HDeoptimize since there may be HParallelMove's
(and other instructions inserted by the following optimization passes) at the
end of the loop pre-header.

Bug: 21034044
Change-Id: I2b6de1f22fbc4568ca419f76382ebd87806d9694