[WIP] JDWP: make asynchronous invoke
... not ready or review ...

When handling an invoke command requested by the debugger, the JDWP
thread used to wait for the event thread to complete the invoke
being able to process another command. Unfortunately, this does not
work well if another thread suspend that event thread and wait for
being resumed by the debugger. Because the JDWP thread is blocked,
it will not process the resume command and cause a deadlock in the
runtime.

This CL changes the way invoke are handled. Now, the JDWP thread will
signal (and resume) the event thread that an invoke is requested. But
it will no longer wait for the invoke to complete and will just keep
processing other commands, without sending a reply for the invoke.
When the event thread completes the invoke, it will send the reply to
the debugger directly (like event threads send event command for
breakpoint, single-step, ...).

Bug: 21515842
Change-Id: I15e00fb068340f3d69dc9225d8d2065246e68c58