WIP: Compress symbols in ART tests.
Quick prototype for anyone who wants to test the compression.
It requires the corresponding decompression in libunwind.
It only works on host, since we do not have the necessary
command line tools on the target.

The script separates .debug_frame .symtab and .symstr to
separate ELF file, compresses it, and stores it in the
main ELF file.

Change-Id: Ib14ab8b7e44122988062880697f18b84946ca38f