[WIP] ART: Single-frame deopt
Add deoptimization of a single frame. Works by removing the managed
code frame and jumping into the quick-to-interpreter bridge, and
the bridge understanding a stored ShadowFrame.

We need a separate fixup pass. For x86, we leave the return address
on the stack so we don't need to push it there.

For Mips64, we move the dispatch address to T9. RA needs to be
preserved, and T9 needs to be set to the function address for actual
runtime functions, anyways.

Bug: 21611912
Change-Id: I06625685ced8b054244f8685ab50b238a705b9d2