[WIP] ART: Single-frame deopt
Add deoptimization of a single frame. Works by removing the managed
code frame and jumping into the quick-to-interpreter bridge, and
the bridge understanding a stored ShadowFrame.

Works for x86, but code needs more cleanup and may go through more
structural changes.

We need a separate fixup pass. For historical reasons, we mark PC
as the register that's being spilled for LR, so contexts actually
load the LR value into the PC slot, which we need to copy over. This
must be done just before the long-jump, as it clobbers values in the
stack and so any stack visitor fails.

Bug: 21611912
Change-Id: I06625685ced8b054244f8685ab50b238a705b9d2