[WIP] ART: Single-frame deopt
Add deoptimization of a single frame. Works by removing the managed
code frame and jumping into the quick-to-interpreter bridge, and
the bridge understanding a stored ShadowFrame.

Works for x86, but code is ugly and may go through structural changes.
Untested for ARM, but hopefully works with LR being always callee-save.

Bug: 21611912
Change-Id: I06625685ced8b054244f8685ab50b238a705b9d2